+++ EDLexerTest.m
@@ -25,7 +25,7 @@
 	NSUInteger i;
 	for (i = 0; i < 3; ++i) {
 		EDLexicalToken *tok = [result.tokens objectAtIndex:i];
-		GHAssertEquals((NSUInteger) EDUnmatchedCharacterToken, tok.type, @"Type should be EDUnmatchedCharacterToken");
+		GHAssertEquals(EDUnmatchedCharacterToken, tok.type, @"Type should be EDUnmatchedCharacterToken");
 		GHAssertEquals((NSUInteger) i, tok.range.location, @"Range location should match char location");
 		GHAssertEquals((NSUInteger) 1, tok.range.length, @"Range length should always be 1");
 	}
@@ -44,11 +44,11 @@
 	EDLexicalToken *t2 = [result.tokens objectAtIndex:1];
 	EDLexicalToken *t3 = [result.tokens objectAtIndex:2];
 	
-	GHAssertEquals((NSUInteger) EDUnmatchedCharacterToken, t1.type, @"First token should be EDUnmatchedCharacterToken");
-	GHAssertEquals((NSUInteger) EDDefinerKeywordToken, t2.type, @"Second token should be EDDefinerKeywordToken");
+	GHAssertEquals(EDUnmatchedCharacterToken, t1.type, @"First token should be EDUnmatchedCharacterToken");
+	GHAssertEquals(EDDefinerKeywordToken, t2.type, @"Second token should be EDDefinerKeywordToken");
 	GHAssertEquals((NSUInteger) 1, t2.range.location, @"Second token should be at offset 1");
 	GHAssertEquals((NSUInteger) 8, t2.range.length, @"Second token should have a length of 8");
-	GHAssertEquals((NSUInteger) EDUnmatchedCharacterToken, t3.type, @"Third token should be EDUnmatchedCharacterToken");
+	GHAssertEquals(EDUnmatchedCharacterToken, t3.type, @"Third token should be EDUnmatchedCharacterToken");
 }
 
 -(void)testUsesTheLongestMatchIfAmbiguous {
@@ -63,7 +63,7 @@
 	
 	EDLexicalToken *tok = [result.tokens objectAtIndex:0];
 	
-	GHAssertEquals((NSUInteger) EDDefinerKeywordToken, tok.type, @"Token should be EDDefinerKeywordToken");
+	GHAssertEquals(EDDefinerKeywordToken, tok.type, @"Token should be EDDefinerKeywordToken");
 	GHAssertEquals((NSUInteger) 0, tok.range.location, @"Token should be at offset 0");
 	GHAssertEquals((NSUInteger) 8, tok.range.length, @"Token should have a length of 8");
 }
