+++ EDWordListLexRuleTest.m
@@ -30,7 +30,7 @@
 	NSString *source = @"hello zapper, how are you?";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(6, source.length - 6)];
 	
-	GHAssertEquals((NSUInteger) EDKeywordToken, (NSUInteger) tok.type, @"Token type should be EDKeywordToken");
+	GHAssertEquals(EDKeywordToken, tok.type, @"Token type should be EDKeywordToken");
 	GHAssertEquals((NSUInteger) 6, (NSUInteger) tok.range.location, @"Token location should be 6");
 	GHAssertEquals((NSUInteger) 6, (NSUInteger) tok.range.length, @"Token length should be 6");
 }
@@ -41,7 +41,7 @@
 	NSString *source = @"hello zapper, how are you?";
 	EDLexicalToken *tok = [rule lexInString:source range:NSMakeRange(6, source.length - 6)];
 	
-	GHAssertEquals((NSUInteger) EDKeywordToken, (NSUInteger) tok.type, @"Token type should be EDKeywordToken");
+	GHAssertEquals(EDKeywordToken, tok.type, @"Token type should be EDKeywordToken");
 	GHAssertEquals((NSUInteger) 6, (NSUInteger) tok.range.location, @"Token location should be 6");
 	GHAssertEquals((NSUInteger) 6, (NSUInteger) tok.range.length, @"Token length should be 6");
 }
